Update libfsntfs to version 20221023 / rev 17 via SR 1038468
https://build.opensuse.org/request/show/1038468
by user jengelh + dimstar_suse
delete system-libs.patch

+++ packages/libf/libfsntfs/libfsntfs.spec
@@ -27,11 +27,8 @@ URL:            https://github.com/libyal/libfsntfs
 Source:         https://github.com/libyal/libfsntfs/releases/download/%version/libfsntfs-experimental-%version.tar.gz
 Source2:        https://github.com/libyal/libfsntfs/releases/download/%version/libfsntfs-experimental-%version.tar.gz.asc
 Source3:        %name.keyring
-Patch1:         system-libs.patch
 BuildRequires:  %{python_module devel}
 BuildRequires:  c_compiler
-BuildRequires:  gettext-tools >= 0.21
-BuildRequires:  libtool
 BuildRequires:  pkg-config
 BuildRequires:  python-rpm-macros
 BuildRequires:  pkgconfig(fuse) >= 2.6
@@ -53,6 +50,7 @@ BuildRequires:  pkgconfig(libfwnt) >= 20220922
 BuildRequires:  pkgconfig(libhmac) >= 20220425
 BuildRequires:  pkgconfig(libuna) >= 20220611
 %python_subpackages
+# Various notes: https://en.opensuse.org/libyal
 
 %description
 Library and tools to access the New Technology File System (NTFS).
@@ -95,22 +93,19 @@ applications that want to make use of %{name}.
 %autosetup -p1
 
 %build
-autoreconf -fi
-# OOT builds are presently broken, so we have to install
-# within each python iteration now, not in %%install.
 %{python_expand #
-# see libcdata for version-sc
 echo "V_%version { global: *; };" >v.sym
 %configure --disable-static --enable-wide-character-type \
 	--enable-python PYTHON_VERSION="%{$python_bin_suffix}" \
 	LDFLAGS="-Wl,--version-script=$PWD/v.sym"
+grep ' '' ''local' config.log && exit 1
 %make_build
 %make_install DESTDIR="%_builddir/rt"
 %make_build clean
 }
 
 %install
-mv %_builddir/rt/* %buildroot/
+mv "%_builddir/rt"/* %buildroot/
 find %{buildroot} -type f -name "*.la" -delete -print
 
 %post   -n %{lname} -p /sbin/ldconfig
